codegen: update Java model path
