+++ java/lance-namespace-apache-client/.openapi-generator/FILES
@@ -36,18 +36,18 @@ src/main/java/com/lancedb/lance/namespace/client/apache/auth/ApiKeyAuth.java
 src/main/java/com/lancedb/lance/namespace/client/apache/auth/Authentication.java
 src/main/java/com/lancedb/lance/namespace/client/apache/auth/HttpBasicAuth.java
 src/main/java/com/lancedb/lance/namespace/client/apache/auth/HttpBearerAuth.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/CreateNamespaceRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/DropNamespaceRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/ErrorResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/GetNamespaceRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/GetNamespaceResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/GetTableRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/GetTableResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/ListNamespacesRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/ListNamespacesResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/NamespaceExistsRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/NamespaceExistsResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/RegisterTableRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/RegisterTableResponse.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/TableExistsRequest.java
-src/main/java/com/lancedb/lance/namespace/client/apache/model/TableExistsResponse.java
+src/main/java/com/lancedb/lance/namespace/model/CreateNamespaceRequest.java
+src/main/java/com/lancedb/lance/namespace/model/DropNamespaceRequest.java
+src/main/java/com/lancedb/lance/namespace/model/ErrorResponse.java
+src/main/java/com/lancedb/lance/namespace/model/GetNamespaceRequest.java
+src/main/java/com/lancedb/lance/namespace/model/GetNamespaceResponse.java
+src/main/java/com/lancedb/lance/namespace/model/GetTableRequest.java
+src/main/java/com/lancedb/lance/namespace/model/GetTableResponse.java
+src/main/java/com/lancedb/lance/namespace/model/ListNamespacesRequest.java
+src/main/java/com/lancedb/lance/namespace/model/ListNamespacesResponse.java
+src/main/java/com/lancedb/lance/namespace/model/NamespaceExistsRequest.java
+src/main/java/com/lancedb/lance/namespace/model/NamespaceExistsResponse.java
+src/main/java/com/lancedb/lance/namespace/model/RegisterTableRequest.java
+src/main/java/com/lancedb/lance/namespace/model/RegisterTableResponse.java
+src/main/java/com/lancedb/lance/namespace/model/TableExistsRequest.java
+src/main/java/com/lancedb/lance/namespace/model/TableExistsResponse.java
