+++ java/lance-namespace-apache-client/src/main/java/com/lancedb/lance/namespace/model/CreateNamespaceRequest.java
@@ -11,11 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lancedb.lance.namespace.client.apache.model;
+package com.lancedb.lance.namespace.model;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.annotation.JsonValue;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
@@ -26,32 +28,67 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.StringJoiner;
 
-/** RegisterTableResponse */
+/** CreateNamespaceRequest */
 @JsonPropertyOrder({
-  RegisterTableResponse.JSON_PROPERTY_NAME,
-  RegisterTableResponse.JSON_PROPERTY_NAMESPACE,
-  RegisterTableResponse.JSON_PROPERTY_LOCATION,
-  RegisterTableResponse.JSON_PROPERTY_PROPERTIES
+  CreateNamespaceRequest.JSON_PROPERTY_NAME,
+  CreateNamespaceRequest.JSON_PROPERTY_PARENT,
+  CreateNamespaceRequest.JSON_PROPERTY_MODE,
+  CreateNamespaceRequest.JSON_PROPERTY_OPTIONS
 })
 @javax.annotation.Generated(
     value = "org.openapitools.codegen.languages.JavaClientCodegen",
     comments = "Generator version: 7.12.0")
-public class RegisterTableResponse {
+public class CreateNamespaceRequest {
   public static final String JSON_PROPERTY_NAME = "name";
   @javax.annotation.Nonnull private String name;
 
-  public static final String JSON_PROPERTY_NAMESPACE = "namespace";
-  @javax.annotation.Nonnull private List<String> namespace = new ArrayList<>();
+  public static final String JSON_PROPERTY_PARENT = "parent";
+  @javax.annotation.Nullable private List<String> parent = new ArrayList<>();
 
-  public static final String JSON_PROPERTY_LOCATION = "location";
-  @javax.annotation.Nonnull private String location;
+  /** Gets or Sets mode */
+  public enum ModeEnum {
+    CREATE(String.valueOf("CREATE")),
 
-  public static final String JSON_PROPERTY_PROPERTIES = "properties";
-  @javax.annotation.Nullable private Map<String, String> properties = new HashMap<>();
+    EXIST_OK(String.valueOf("EXIST_OK")),
 
-  public RegisterTableResponse() {}
+    OVERWRITE(String.valueOf("OVERWRITE"));
 
-  public RegisterTableResponse name(@javax.annotation.Nonnull String name) {
+    private String value;
+
+    ModeEnum(String value) {
+      this.value = value;
+    }
+
+    @JsonValue
+    public String getValue() {
+      return value;
+    }
+
+    @Override
+    public String toString() {
+      return String.valueOf(value);
+    }
+
+    @JsonCreator
+    public static ModeEnum fromValue(String value) {
+      for (ModeEnum b : ModeEnum.values()) {
+        if (b.value.equals(value)) {
+          return b;
+        }
+      }
+      throw new IllegalArgumentException("Unexpected value '" + value + "'");
+    }
+  }
+
+  public static final String JSON_PROPERTY_MODE = "mode";
+  @javax.annotation.Nonnull private ModeEnum mode;
+
+  public static final String JSON_PROPERTY_OPTIONS = "options";
+  @javax.annotation.Nullable private Map<String, String> options = new HashMap<>();
+
+  public CreateNamespaceRequest() {}
+
+  public CreateNamespaceRequest name(@javax.annotation.Nonnull String name) {
 
     this.name = name;
     return this;
@@ -75,93 +112,92 @@ public class RegisterTableResponse {
     this.name = name;
   }
 
-  public RegisterTableResponse namespace(@javax.annotation.Nonnull List<String> namespace) {
+  public CreateNamespaceRequest parent(@javax.annotation.Nullable List<String> parent) {
 
-    this.namespace = namespace;
+    this.parent = parent;
     return this;
   }
 
-  public RegisterTableResponse addNamespaceItem(String namespaceItem) {
-    if (this.namespace == null) {
-      this.namespace = new ArrayList<>();
+  public CreateNamespaceRequest addParentItem(String parentItem) {
+    if (this.parent == null) {
+      this.parent = new ArrayList<>();
     }
-    this.namespace.add(namespaceItem);
+    this.parent.add(parentItem);
     return this;
   }
 
   /**
-   * Get namespace
+   * Get parent
    *
-   * @return namespace
-   */
-  @javax.annotation.Nonnull
-  @JsonProperty(JSON_PROPERTY_NAMESPACE)
-  @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public List<String> getNamespace() {
-    return namespace;
-  }
-
-  @JsonProperty(JSON_PROPERTY_NAMESPACE)
-  @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setNamespace(@javax.annotation.Nonnull List<String> namespace) {
-    this.namespace = namespace;
-  }
-
-  public RegisterTableResponse location(@javax.annotation.Nonnull String location) {
-
-    this.location = location;
-    return this;
-  }
-
-  /**
-   * Get location
-   *
-   * @return location
-   */
-  @javax.annotation.Nonnull
-  @JsonProperty(JSON_PROPERTY_LOCATION)
-  @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public String getLocation() {
-    return location;
-  }
-
-  @JsonProperty(JSON_PROPERTY_LOCATION)
-  @JsonInclude(value = JsonInclude.Include.ALWAYS)
-  public void setLocation(@javax.annotation.Nonnull String location) {
-    this.location = location;
-  }
-
-  public RegisterTableResponse properties(
-      @javax.annotation.Nullable Map<String, String> properties) {
-
-    this.properties = properties;
-    return this;
-  }
-
-  public RegisterTableResponse putPropertiesItem(String key, String propertiesItem) {
-    if (this.properties == null) {
-      this.properties = new HashMap<>();
-    }
-    this.properties.put(key, propertiesItem);
-    return this;
-  }
-
-  /**
-   * Get properties
-   *
-   * @return properties
+   * @return parent
    */
   @javax.annotation.Nullable
-  @JsonProperty(JSON_PROPERTY_PROPERTIES)
+  @JsonProperty(JSON_PROPERTY_PARENT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public Map<String, String> getProperties() {
-    return properties;
+  public List<String> getParent() {
+    return parent;
   }
 
-  @JsonProperty(JSON_PROPERTY_PROPERTIES)
+  @JsonProperty(JSON_PROPERTY_PARENT)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setProperties(@javax.annotation.Nullable Map<String, String> properties) {
-    this.properties = properties;
+  public void setParent(@javax.annotation.Nullable List<String> parent) {
+    this.parent = parent;
+  }
+
+  public CreateNamespaceRequest mode(@javax.annotation.Nonnull ModeEnum mode) {
+
+    this.mode = mode;
+    return this;
+  }
+
+  /**
+   * Get mode
+   *
+   * @return mode
+   */
+  @javax.annotation.Nonnull
+  @JsonProperty(JSON_PROPERTY_MODE)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public ModeEnum getMode() {
+    return mode;
+  }
+
+  @JsonProperty(JSON_PROPERTY_MODE)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public void setMode(@javax.annotation.Nonnull ModeEnum mode) {
+    this.mode = mode;
+  }
+
+  public CreateNamespaceRequest options(@javax.annotation.Nullable Map<String, String> options) {
+
+    this.options = options;
+    return this;
+  }
+
+  public CreateNamespaceRequest putOptionsItem(String key, String optionsItem) {
+    if (this.options == null) {
+      this.options = new HashMap<>();
+    }
+    this.options.put(key, optionsItem);
+    return this;
+  }
+
+  /**
+   * Get options
+   *
+   * @return options
+   */
+  @javax.annotation.Nullable
+  @JsonProperty(JSON_PROPERTY_OPTIONS)
+  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
+  public Map<String, String> getOptions() {
+    return options;
+  }
+
+  @JsonProperty(JSON_PROPERTY_OPTIONS)
+  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
+  public void setOptions(@javax.annotation.Nullable Map<String, String> options) {
+    this.options = options;
   }
 
   @Override
@@ -172,26 +208,26 @@ public class RegisterTableResponse {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    RegisterTableResponse registerTableResponse = (RegisterTableResponse) o;
-    return Objects.equals(this.name, registerTableResponse.name)
-        && Objects.equals(this.namespace, registerTableResponse.namespace)
-        && Objects.equals(this.location, registerTableResponse.location)
-        && Objects.equals(this.properties, registerTableResponse.properties);
+    CreateNamespaceRequest createNamespaceRequest = (CreateNamespaceRequest) o;
+    return Objects.equals(this.name, createNamespaceRequest.name)
+        && Objects.equals(this.parent, createNamespaceRequest.parent)
+        && Objects.equals(this.mode, createNamespaceRequest.mode)
+        && Objects.equals(this.options, createNamespaceRequest.options);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(name, namespace, location, properties);
+    return Objects.hash(name, parent, mode, options);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class RegisterTableResponse {\n");
+    sb.append("class CreateNamespaceRequest {\n");
     sb.append("    name: ").append(toIndentedString(name)).append("\n");
-    sb.append("    namespace: ").append(toIndentedString(namespace)).append("\n");
-    sb.append("    location: ").append(toIndentedString(location)).append("\n");
-    sb.append("    properties: ").append(toIndentedString(properties)).append("\n");
+    sb.append("    parent: ").append(toIndentedString(parent)).append("\n");
+    sb.append("    mode: ").append(toIndentedString(mode)).append("\n");
+    sb.append("    options: ").append(toIndentedString(options)).append("\n");
     sb.append("}");
     return sb.toString();
   }
@@ -253,19 +289,19 @@ public class RegisterTableResponse {
       }
     }
 
-    // add `namespace` to the URL query string
-    if (getNamespace() != null) {
-      for (int i = 0; i < getNamespace().size(); i++) {
+    // add `parent` to the URL query string
+    if (getParent() != null) {
+      for (int i = 0; i < getParent().size(); i++) {
         try {
           joiner.add(
               String.format(
-                  "%snamespace%s%s=%s",
+                  "%sparent%s%s=%s",
                   prefix,
                   suffix,
                   "".equals(suffix)
                       ? ""
                       : String.format("%s%d%s", containerPrefix, i, containerSuffix),
-                  URLEncoder.encode(String.valueOf(getNamespace().get(i)), "UTF-8")
+                  URLEncoder.encode(String.valueOf(getParent().get(i)), "UTF-8")
                       .replaceAll("\\+", "%20")));
         } catch (UnsupportedEncodingException e) {
           // Should never happen, UTF-8 is always supported
@@ -274,36 +310,35 @@ public class RegisterTableResponse {
       }
     }
 
-    // add `location` to the URL query string
-    if (getLocation() != null) {
+    // add `mode` to the URL query string
+    if (getMode() != null) {
       try {
         joiner.add(
             String.format(
-                "%slocation%s=%s",
+                "%smode%s=%s",
                 prefix,
                 suffix,
-                URLEncoder.encode(String.valueOf(getLocation()), "UTF-8")
-                    .replaceAll("\\+", "%20")));
+                URLEncoder.encode(String.valueOf(getMode()), "UTF-8").replaceAll("\\+", "%20")));
       } catch (UnsupportedEncodingException e) {
         // Should never happen, UTF-8 is always supported
         throw new RuntimeException(e);
       }
     }
 
-    // add `properties` to the URL query string
-    if (getProperties() != null) {
-      for (String _key : getProperties().keySet()) {
+    // add `options` to the URL query string
+    if (getOptions() != null) {
+      for (String _key : getOptions().keySet()) {
         try {
           joiner.add(
               String.format(
-                  "%sproperties%s%s=%s",
+                  "%soptions%s%s=%s",
                   prefix,
                   suffix,
                   "".equals(suffix)
                       ? ""
                       : String.format("%s%d%s", containerPrefix, _key, containerSuffix),
-                  getProperties().get(_key),
-                  URLEncoder.encode(String.valueOf(getProperties().get(_key)), "UTF-8")
+                  getOptions().get(_key),
+                  URLEncoder.encode(String.valueOf(getOptions().get(_key)), "UTF-8")
                       .replaceAll("\\+", "%20")));
         } catch (UnsupportedEncodingException e) {
           // Should never happen, UTF-8 is always supported
