+++ java/lance-namespace-apache-client/src/main/java/com/lancedb/lance/namespace/model/GetTableResponse.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lancedb.lance.namespace.client.apache.model;
+package com.lancedb.lance.namespace.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -20,25 +20,38 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.StringJoiner;
 
-/** GetTableRequest */
-@JsonPropertyOrder({GetTableRequest.JSON_PROPERTY_NAME, GetTableRequest.JSON_PROPERTY_NAMESPACE})
+/** GetTableResponse */
+@JsonPropertyOrder({
+  GetTableResponse.JSON_PROPERTY_NAME,
+  GetTableResponse.JSON_PROPERTY_NAMESPACE,
+  GetTableResponse.JSON_PROPERTY_LOCATION,
+  GetTableResponse.JSON_PROPERTY_PROPERTIES
+})
 @javax.annotation.Generated(
     value = "org.openapitools.codegen.languages.JavaClientCodegen",
     comments = "Generator version: 7.12.0")
-public class GetTableRequest {
+public class GetTableResponse {
   public static final String JSON_PROPERTY_NAME = "name";
   @javax.annotation.Nonnull private String name;
 
   public static final String JSON_PROPERTY_NAMESPACE = "namespace";
   @javax.annotation.Nonnull private List<String> namespace = new ArrayList<>();
 
-  public GetTableRequest() {}
+  public static final String JSON_PROPERTY_LOCATION = "location";
+  @javax.annotation.Nonnull private String location;
 
-  public GetTableRequest name(@javax.annotation.Nonnull String name) {
+  public static final String JSON_PROPERTY_PROPERTIES = "properties";
+  @javax.annotation.Nullable private Map<String, String> properties = new HashMap<>();
+
+  public GetTableResponse() {}
+
+  public GetTableResponse name(@javax.annotation.Nonnull String name) {
 
     this.name = name;
     return this;
@@ -62,13 +75,13 @@ public class GetTableRequest {
     this.name = name;
   }
 
-  public GetTableRequest namespace(@javax.annotation.Nonnull List<String> namespace) {
+  public GetTableResponse namespace(@javax.annotation.Nonnull List<String> namespace) {
 
     this.namespace = namespace;
     return this;
   }
 
-  public GetTableRequest addNamespaceItem(String namespaceItem) {
+  public GetTableResponse addNamespaceItem(String namespaceItem) {
     if (this.namespace == null) {
       this.namespace = new ArrayList<>();
     }
@@ -94,6 +107,62 @@ public class GetTableRequest {
     this.namespace = namespace;
   }
 
+  public GetTableResponse location(@javax.annotation.Nonnull String location) {
+
+    this.location = location;
+    return this;
+  }
+
+  /**
+   * Get location
+   *
+   * @return location
+   */
+  @javax.annotation.Nonnull
+  @JsonProperty(JSON_PROPERTY_LOCATION)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public String getLocation() {
+    return location;
+  }
+
+  @JsonProperty(JSON_PROPERTY_LOCATION)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public void setLocation(@javax.annotation.Nonnull String location) {
+    this.location = location;
+  }
+
+  public GetTableResponse properties(@javax.annotation.Nullable Map<String, String> properties) {
+
+    this.properties = properties;
+    return this;
+  }
+
+  public GetTableResponse putPropertiesItem(String key, String propertiesItem) {
+    if (this.properties == null) {
+      this.properties = new HashMap<>();
+    }
+    this.properties.put(key, propertiesItem);
+    return this;
+  }
+
+  /**
+   * Get properties
+   *
+   * @return properties
+   */
+  @javax.annotation.Nullable
+  @JsonProperty(JSON_PROPERTY_PROPERTIES)
+  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
+  public Map<String, String> getProperties() {
+    return properties;
+  }
+
+  @JsonProperty(JSON_PROPERTY_PROPERTIES)
+  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
+  public void setProperties(@javax.annotation.Nullable Map<String, String> properties) {
+    this.properties = properties;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -102,22 +171,26 @@ public class GetTableRequest {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    GetTableRequest getTableRequest = (GetTableRequest) o;
-    return Objects.equals(this.name, getTableRequest.name)
-        && Objects.equals(this.namespace, getTableRequest.namespace);
+    GetTableResponse getTableResponse = (GetTableResponse) o;
+    return Objects.equals(this.name, getTableResponse.name)
+        && Objects.equals(this.namespace, getTableResponse.namespace)
+        && Objects.equals(this.location, getTableResponse.location)
+        && Objects.equals(this.properties, getTableResponse.properties);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(name, namespace);
+    return Objects.hash(name, namespace, location, properties);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class GetTableRequest {\n");
+    sb.append("class GetTableResponse {\n");
     sb.append("    name: ").append(toIndentedString(name)).append("\n");
     sb.append("    namespace: ").append(toIndentedString(namespace)).append("\n");
+    sb.append("    location: ").append(toIndentedString(location)).append("\n");
+    sb.append("    properties: ").append(toIndentedString(properties)).append("\n");
     sb.append("}");
     return sb.toString();
   }
@@ -192,6 +265,44 @@ public class GetTableRequest {
                       ? ""
                       : String.format("%s%d%s", containerPrefix, i, containerSuffix),
                   URLEncoder.encode(String.valueOf(getNamespace().get(i)), "UTF-8")
+                      .replaceAll("\\+", "%20")));
+        } catch (UnsupportedEncodingException e) {
+          // Should never happen, UTF-8 is always supported
+          throw new RuntimeException(e);
+        }
+      }
+    }
+
+    // add `location` to the URL query string
+    if (getLocation() != null) {
+      try {
+        joiner.add(
+            String.format(
+                "%slocation%s=%s",
+                prefix,
+                suffix,
+                URLEncoder.encode(String.valueOf(getLocation()), "UTF-8")
+                    .replaceAll("\\+", "%20")));
+      } catch (UnsupportedEncodingException e) {
+        // Should never happen, UTF-8 is always supported
+        throw new RuntimeException(e);
+      }
+    }
+
+    // add `properties` to the URL query string
+    if (getProperties() != null) {
+      for (String _key : getProperties().keySet()) {
+        try {
+          joiner.add(
+              String.format(
+                  "%sproperties%s%s=%s",
+                  prefix,
+                  suffix,
+                  "".equals(suffix)
+                      ? ""
+                      : String.format("%s%d%s", containerPrefix, _key, containerSuffix),
+                  getProperties().get(_key),
+                  URLEncoder.encode(String.valueOf(getProperties().get(_key)), "UTF-8")
                       .replaceAll("\\+", "%20")));
         } catch (UnsupportedEncodingException e) {
           // Should never happen, UTF-8 is always supported
