+++ java/lance-namespace-apache-client/src/main/java/com/lancedb/lance/namespace/model/ListNamespacesResponse.java
@@ -11,75 +11,73 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lancedb.lance.namespace.client.apache.model;
+package com.lancedb.lance.namespace.model;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.LinkedHashSet;
 import java.util.Objects;
+import java.util.Set;
 import java.util.StringJoiner;
 
-/** ListNamespacesRequest */
+/** ListNamespacesResponse */
 @JsonPropertyOrder({
-  ListNamespacesRequest.JSON_PROPERTY_PARENT,
-  ListNamespacesRequest.JSON_PROPERTY_PAGE_TOKEN,
-  ListNamespacesRequest.JSON_PROPERTY_PAGE_SIZE
+  ListNamespacesResponse.JSON_PROPERTY_NAMESPACES,
+  ListNamespacesResponse.JSON_PROPERTY_NEXT_PAGE_TOKEN
 })
 @javax.annotation.Generated(
     value = "org.openapitools.codegen.languages.JavaClientCodegen",
     comments = "Generator version: 7.12.0")
-public class ListNamespacesRequest {
-  public static final String JSON_PROPERTY_PARENT = "parent";
-  @javax.annotation.Nullable private List<String> parent = new ArrayList<>();
+public class ListNamespacesResponse {
+  public static final String JSON_PROPERTY_NAMESPACES = "namespaces";
+  @javax.annotation.Nonnull private Set<String> namespaces = new LinkedHashSet<>();
 
-  public static final String JSON_PROPERTY_PAGE_TOKEN = "pageToken";
-  @javax.annotation.Nullable private String pageToken;
+  public static final String JSON_PROPERTY_NEXT_PAGE_TOKEN = "nextPageToken";
+  @javax.annotation.Nullable private String nextPageToken;
 
-  public static final String JSON_PROPERTY_PAGE_SIZE = "pageSize";
-  @javax.annotation.Nullable private Integer pageSize;
+  public ListNamespacesResponse() {}
 
-  public ListNamespacesRequest() {}
+  public ListNamespacesResponse namespaces(@javax.annotation.Nonnull Set<String> namespaces) {
 
-  public ListNamespacesRequest parent(@javax.annotation.Nullable List<String> parent) {
-
-    this.parent = parent;
+    this.namespaces = namespaces;
     return this;
   }
 
-  public ListNamespacesRequest addParentItem(String parentItem) {
-    if (this.parent == null) {
-      this.parent = new ArrayList<>();
+  public ListNamespacesResponse addNamespacesItem(String namespacesItem) {
+    if (this.namespaces == null) {
+      this.namespaces = new LinkedHashSet<>();
     }
-    this.parent.add(parentItem);
+    this.namespaces.add(namespacesItem);
     return this;
   }
 
   /**
-   * Get parent
+   * Get namespaces
    *
-   * @return parent
+   * @return namespaces
    */
-  @javax.annotation.Nullable
-  @JsonProperty(JSON_PROPERTY_PARENT)
-  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public List<String> getParent() {
-    return parent;
+  @javax.annotation.Nonnull
+  @JsonProperty(JSON_PROPERTY_NAMESPACES)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public Set<String> getNamespaces() {
+    return namespaces;
   }
 
-  @JsonProperty(JSON_PROPERTY_PARENT)
-  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setParent(@javax.annotation.Nullable List<String> parent) {
-    this.parent = parent;
+  @JsonDeserialize(as = LinkedHashSet.class)
+  @JsonProperty(JSON_PROPERTY_NAMESPACES)
+  @JsonInclude(value = JsonInclude.Include.ALWAYS)
+  public void setNamespaces(@javax.annotation.Nonnull Set<String> namespaces) {
+    this.namespaces = namespaces;
   }
 
-  public ListNamespacesRequest pageToken(@javax.annotation.Nullable String pageToken) {
+  public ListNamespacesResponse nextPageToken(@javax.annotation.Nullable String nextPageToken) {
 
-    this.pageToken = pageToken;
+    this.nextPageToken = nextPageToken;
     return this;
   }
 
@@ -93,43 +91,19 @@ public class ListNamespacesRequest {
    * Clients must interpret either &#x60;null&#x60;, missing value or empty string value of
    * &#x60;nextPageToken&#x60; from a server response as the end of the listing results.
    *
-   * @return pageToken
+   * @return nextPageToken
    */
   @javax.annotation.Nullable
-  @JsonProperty(JSON_PROPERTY_PAGE_TOKEN)
+  @JsonProperty(JSON_PROPERTY_NEXT_PAGE_TOKEN)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public String getPageToken() {
-    return pageToken;
+  public String getNextPageToken() {
+    return nextPageToken;
   }
 
-  @JsonProperty(JSON_PROPERTY_PAGE_TOKEN)
+  @JsonProperty(JSON_PROPERTY_NEXT_PAGE_TOKEN)
   @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setPageToken(@javax.annotation.Nullable String pageToken) {
-    this.pageToken = pageToken;
-  }
-
-  public ListNamespacesRequest pageSize(@javax.annotation.Nullable Integer pageSize) {
-
-    this.pageSize = pageSize;
-    return this;
-  }
-
-  /**
-   * An inclusive upper bound of the number of results that a client will receive.
-   *
-   * @return pageSize
-   */
-  @javax.annotation.Nullable
-  @JsonProperty(JSON_PROPERTY_PAGE_SIZE)
-  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public Integer getPageSize() {
-    return pageSize;
-  }
-
-  @JsonProperty(JSON_PROPERTY_PAGE_SIZE)
-  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
-  public void setPageSize(@javax.annotation.Nullable Integer pageSize) {
-    this.pageSize = pageSize;
+  public void setNextPageToken(@javax.annotation.Nullable String nextPageToken) {
+    this.nextPageToken = nextPageToken;
   }
 
   @Override
@@ -140,24 +114,22 @@ public class ListNamespacesRequest {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    ListNamespacesRequest listNamespacesRequest = (ListNamespacesRequest) o;
-    return Objects.equals(this.parent, listNamespacesRequest.parent)
-        && Objects.equals(this.pageToken, listNamespacesRequest.pageToken)
-        && Objects.equals(this.pageSize, listNamespacesRequest.pageSize);
+    ListNamespacesResponse listNamespacesResponse = (ListNamespacesResponse) o;
+    return Objects.equals(this.namespaces, listNamespacesResponse.namespaces)
+        && Objects.equals(this.nextPageToken, listNamespacesResponse.nextPageToken);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(parent, pageToken, pageSize);
+    return Objects.hash(namespaces, nextPageToken);
   }
 
   @Override
   public String toString() {
     StringBuilder sb = new StringBuilder();
-    sb.append("class ListNamespacesRequest {\n");
-    sb.append("    parent: ").append(toIndentedString(parent)).append("\n");
-    sb.append("    pageToken: ").append(toIndentedString(pageToken)).append("\n");
-    sb.append("    pageSize: ").append(toIndentedString(pageSize)).append("\n");
+    sb.append("class ListNamespacesResponse {\n");
+    sb.append("    namespaces: ").append(toIndentedString(namespaces)).append("\n");
+    sb.append("    nextPageToken: ").append(toIndentedString(nextPageToken)).append("\n");
     sb.append("}");
     return sb.toString();
   }
@@ -204,52 +176,37 @@ public class ListNamespacesRequest {
 
     StringJoiner joiner = new StringJoiner("&");
 
-    // add `parent` to the URL query string
-    if (getParent() != null) {
-      for (int i = 0; i < getParent().size(); i++) {
+    // add `namespaces` to the URL query string
+    if (getNamespaces() != null) {
+      int i = 0;
+      for (String _item : getNamespaces()) {
         try {
           joiner.add(
               String.format(
-                  "%sparent%s%s=%s",
+                  "%snamespaces%s%s=%s",
                   prefix,
                   suffix,
                   "".equals(suffix)
                       ? ""
                       : String.format("%s%d%s", containerPrefix, i, containerSuffix),
-                  URLEncoder.encode(String.valueOf(getParent().get(i)), "UTF-8")
-                      .replaceAll("\\+", "%20")));
+                  URLEncoder.encode(String.valueOf(_item), "UTF-8").replaceAll("\\+", "%20")));
         } catch (UnsupportedEncodingException e) {
           // Should never happen, UTF-8 is always supported
           throw new RuntimeException(e);
         }
       }
+      i++;
     }
 
-    // add `pageToken` to the URL query string
-    if (getPageToken() != null) {
+    // add `nextPageToken` to the URL query string
+    if (getNextPageToken() != null) {
       try {
         joiner.add(
             String.format(
-                "%spageToken%s=%s",
+                "%snextPageToken%s=%s",
                 prefix,
                 suffix,
-                URLEncoder.encode(String.valueOf(getPageToken()), "UTF-8")
-                    .replaceAll("\\+", "%20")));
-      } catch (UnsupportedEncodingException e) {
-        // Should never happen, UTF-8 is always supported
-        throw new RuntimeException(e);
-      }
-    }
-
-    // add `pageSize` to the URL query string
-    if (getPageSize() != null) {
-      try {
-        joiner.add(
-            String.format(
-                "%spageSize%s=%s",
-                prefix,
-                suffix,
-                URLEncoder.encode(String.valueOf(getPageSize()), "UTF-8")
+                URLEncoder.encode(String.valueOf(getNextPageToken()), "UTF-8")
                     .replaceAll("\\+", "%20")));
       } catch (UnsupportedEncodingException e) {
         // Should never happen, UTF-8 is always supported
